fixed some errors after merge

+++ tests/branching_kernels/if_else_in_if_and_else_of_if_else_gfx/if_else_in_if_and_else_of_if_else_gfx_dcmpl-gfx1010.cl
@@ -31,10 +31,10 @@ void if_else_in_if_and_else_of_if_else_2(int x, __global int *data, int y)
     if (1 != get_global_id(0)) {
         data[get_global_id(0)] = get_global_id(0) * y;
         if (get_global_id(1) != get_global_id(2)) {
-            var5 = (ulong)y + (ulong)get_global_id(1);
+            var5 = (ulong)y + get_global_id(1);
         }
         else {
-            var5 = (ulong)x + (ulong)get_global_id(1);
+            var5 = (ulong)x + get_global_id(1);
         }
     }
     else {
